requirements.txt
pandas
scikit-learn
kafka-python
streamlit
streamlit-autorefresh